# #5140: Tags merge optimization