@mstyura 2024-09-12
Updated benchmark results in PR description to latest master (+ extra benchmark) + proposed changes.